+++ components/index.js
@@ -1,0 +1,2 @@
+export { default as ButtonControl } from './ButtonControl';
+export { default as ScheduleCard } from './ScheduleCard';
