+++ screens/Event.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Text, View } from 'react-native';
 
-export default class Event extends Component<Props> {    
+class Event extends Component<Props> {
     render() {
         const event = this.props.navigation.getParam('event', '');
 
@@ -21,3 +21,5 @@ export default class Event extends Component<Props> {
         )
     }
 }
+
+export default Event;
