+++ screens/Home.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Text, View, StyleSheet } from 'react-native';
 
-export default class Home extends Component<Props> {
+class Home extends Component<Props> {
   static navigationOptions = {
     title: 'Home',
     headerLeft: null
@@ -57,3 +57,5 @@ const styles = StyleSheet.create({
     marginTop: 25,
   }
 });
+
+export default Home;
