+++ screens/index.js
@@ -1,0 +1,4 @@
+export { default as Event } from './Event';
+export { default as Home } from './Home';
+export { default as Notifications } from './Notifications';
+export { default as Schedule } from './Schedule';
